Update patch set 1
Patch Set 1:

I think Liam that we should release the simulator next time we do a release, may be we can can do a full or partial release towards the end of June, but moving the version forward instead of backwards.   I am a little uneasy about moving back the release version even though the change may be short lived.   I think it is OK in the meantime to operate with snapshots for it (we could regenerate them with "remerge" if they go away in nexus3) unless there is a pressing need to use a released version right away.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>