Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think Liam that we should release the simulator next time we do a release, may be we can can do a full or partial release towards the end of June, but moving the version forward instead of backwards.   I am a little uneasy about moving back the release version even though the change may be short lived.   I think it is OK in the meantime to operate with snapshots for it (we could regenerate them with "remerge" if they go away in nexus3) unless there is a pressing need to use a released version right away.

I am going to step the version of policy models but I have to step the versions of common and the parent PM back to the released versions of those.

So we will release policy-models to version 2.4.4, but because we are not releasing common and policy parent, we need to go back to the released versions of those.

We have to release policy models because otherwise we can't get the CSIT tests working in Jakarta/Istanbul/Honolulu.

So the steps in the Jakarta branch are:
1. Step back the versions of policy parent and policy common referred to by policy models to their released versions (This commit)
2. Release policy models version 2.4.4 maven artifacts
3. Release policy models-simulator version 2.4.4
4. Step the snapshot version of poliocy-models in jakarta to 2.4.5-SNAPSHOT and reset the referred versions of policy-parent and policy-common to 3.5.4-SNAPSHOT and 1.10.4-SNAPSHOT respectively.

I don't think we need to change any components that use policy-models to version 2.4.4 as 2.4.4 is teh same as 2.4.3 except that models-simulator is released.

Patch-set: 1
Reviewer: Gerrit User 2787 <2787@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>