Update patch set 1
Patch Set 1:

The SON Use Case Team requested to have Policy talking directory to A1-PMS.   This is to accommodate that request.   I would like to put it in Kohn if folks agree to help the use case out.   There is a companion drools-apps review that I will put out once this is merged.

Patch-set: 1